+++ dimas/src/com/liveliness_builder.rs
@@ -8,7 +8,8 @@ use super::liveliness::{ArcLivelinessCallback, LivelinessSubscriber};
 use dimas_core::{
 	enums::OperationState,
 	error::{DimasError, Result},
-	traits::Context, utils::selector_from,
+	traits::Context,
+	utils::selector_from,
 };
 use std::{
 	collections::HashMap,